+++ tests/testthat/test-gm_auth.R
@@ -1,12 +1,44 @@
-# https://github.com/r-lib/gmailr/issues/144
-test_that("gm_auth_configure() can accept the app via `app`", {
+test_that("gm_auth_configure() errors for key, secret, appname, app", {
+  expect_snapshot(
+    error = TRUE,
+    gm_auth_configure(key = "KEY", secret = "SECRET")
+  )
+  expect_error(gm_auth_configure(appname = "APPNAME"))
   google_app <- httr::oauth_app(
     "gmailr",
     key = "KEYKEYKEY",
     secret = "SECRETSECRETSECRET"
   )
+  expect_error(gm_auth_configure(app = google_app))
+})
 
-  expect_error_free(gm_auth_configure(app = google_app))
+test_that("gm_oauth_app() is deprecated", {
+  withr::local_options(lifecycle_verbosity = "warning")
+  expect_snapshot(absorb <- gm_oauth_app())
+})
+
+test_that("gm_auth_configure() works", {
+  old_client <- gm_oauth_client()
+  withr::defer(gm_auth_configure(client = old_client))
+
+  expect_snapshot(
+    gm_auth_configure(client = gargle::gargle_client(), path = "PATH"),
+    error = TRUE
+  )
+
+  gm_auth_configure(client = gargle::gargle_client())
+  expect_s3_class(gm_oauth_client(), "gargle_oauth_client")
+
+  gm_auth_configure(
+    path = system.file(
+      "extdata", "client_secret_installed.googleusercontent.com.json",
+      package = "gargle"
+    )
+  )
+  expect_s3_class(gm_oauth_client(), "gargle_oauth_client")
+
+  gm_auth_configure(client = NULL)
+  expect_null(gm_oauth_client())
 })
 
 test_that("gm_scopes() reveals gmail scopes", {